Perform pool ACL evaluation in pool list handler
