+++ src/pool/srv_pool.c
@@ -2907,174 +2907,6 @@ bulk_cb(const struct crt_bulk_cb_info *cb_info)
 	return 0;
 }
 
-int
-ds_pool_check_access(uuid_t pool_uuid, d_rank_list_t *ps_ranks, unsigned int flags, d_iov_t *cred)
-{
-	struct rsvc_client            client;
-	crt_endpoint_t                ep;
-	struct dss_module_info       *info = dss_get_module_info();
-	crt_rpc_t                    *rpc;
-	struct pool_check_access_in  *in;
-	struct pool_check_access_out *out;
-	int                           rc;
-
-	D_DEBUG(DB_MD, DF_UUID ": checking pool access\n", DP_UUID(pool_uuid));
-
-	rc = rsvc_client_init(&client, ps_ranks);
-	if (rc != 0)
-		goto out;
-
-rechoose:
-	ep.ep_grp = NULL; /* primary group */
-	rc        = rsvc_client_choose(&client, &ep);
-	if (rc != 0) {
-		D_ERROR(DF_UUID ": cannot find pool service: " DF_RC "\n", DP_UUID(pool_uuid),
-			DP_RC(rc));
-		goto out_client;
-	}
-
-	rc = pool_req_create(info->dmi_ctx, &ep, POOL_CHECK_ACCESS, &rpc);
-	if (rc != 0) {
-		D_ERROR(DF_UUID ": failed to create pool check access rpc, " DF_RC "\n",
-			DP_UUID(pool_uuid), DP_RC(rc));
-		goto out_client;
-	}
-
-	in             = crt_req_get(rpc);
-	in->pcai_flags = flags;
-	uuid_copy(in->pcai_op.pi_uuid, pool_uuid);
-	uuid_clear(in->pcai_op.pi_hdl);
-
-	rc = daos_iov_copy(&in->pcai_cred, cred);
-	if (rc != 0) {
-		D_ERROR(DF_UUID ": failed to copy credential IOV, " DF_RC "\n", DP_UUID(pool_uuid),
-			DP_RC(rc));
-		goto out_client;
-	}
-
-	rc  = dss_rpc_send(rpc);
-	out = crt_reply_get(rpc);
-	D_ASSERT(out != NULL);
-
-	rc = pool_rsvc_client_complete_rpc(&client, &ep, rc, &out->pcao_op);
-	if (rc == RSVC_CLIENT_RECHOOSE) {
-		crt_req_decref(rpc);
-		dss_sleep(RECHOOSE_SLEEP_MS);
-		goto rechoose;
-	}
-
-	rc = out->pcao_op.po_rc;
-	if (rc != 0 && rc != -DER_NO_PERM) {
-		D_ERROR(DF_UUID ": failed to check pool access, " DF_RC "\n", DP_UUID(pool_uuid),
-			DP_RC(rc));
-	}
-
-	crt_req_decref(rpc);
-out_client:
-	rsvc_client_fini(&client);
-out:
-	return rc;
-}
-
-static int
-pool_check_cred_access(d_iov_t *cred, int flags, daos_prop_t *prop)
-{
-	struct daos_prop_entry *acl_entry;
-	struct d_ownership      owner;
-	struct daos_prop_entry *owner_entry;
-	struct daos_prop_entry *owner_grp_entry;
-	uint64_t                sec_capas = 0;
-	int                     rc;
-
-	D_ASSERT(prop != NULL);
-
-	acl_entry = daos_prop_entry_get(prop, DAOS_PROP_PO_ACL);
-	D_ASSERT(acl_entry != NULL);
-	D_ASSERT(acl_entry->dpe_val_ptr != NULL);
-
-	owner_entry = daos_prop_entry_get(prop, DAOS_PROP_PO_OWNER);
-	D_ASSERT(owner_entry != NULL);
-	D_ASSERT(owner_entry->dpe_str != NULL);
-
-	owner_grp_entry = daos_prop_entry_get(prop, DAOS_PROP_PO_OWNER_GROUP);
-	D_ASSERT(owner_grp_entry != NULL);
-	D_ASSERT(owner_grp_entry->dpe_str != NULL);
-
-	owner.user  = owner_entry->dpe_str;
-	owner.group = owner_grp_entry->dpe_str;
-
-	/*
-	 * Security capabilities determine the access control policy on this
-	 * pool handle.
-	 */
-	rc = ds_sec_pool_get_capabilities(flags, cred, &owner, acl_entry->dpe_val_ptr, &sec_capas);
-	if (rc != 0) {
-		DL_ERROR(rc, "failed to read sec capabilities");
-		return rc;
-	}
-
-	if (!ds_sec_pool_can_connect(sec_capas)) {
-		return -DER_NO_PERM;
-	}
-
-	return 0;
-}
-
-void
-ds_pool_check_access_handler(crt_rpc_t *rpc)
-{
-	struct pool_check_access_in  *in  = crt_req_get(rpc);
-	struct pool_check_access_out *out = crt_reply_get(rpc);
-	struct pool_svc              *svc;
-	struct rdb_tx                 tx;
-	daos_prop_t                  *prop = NULL;
-	uint64_t                      prop_bits;
-	int                           rc;
-
-	D_DEBUG(DB_MD, DF_UUID ": processing rpc: %p hdl=" DF_UUID "\n",
-		DP_UUID(in->pcai_op.pi_uuid), rpc, DP_UUID(in->pcai_op.pi_hdl));
-
-	rc = pool_svc_lookup_leader(in->pcai_op.pi_uuid, &svc, &out->pcao_op.po_hint);
-	if (rc != 0)
-		D_GOTO(out, rc);
-
-	rc = rdb_tx_begin(svc->ps_rsvc.s_db, svc->ps_rsvc.s_term, &tx);
-	if (rc != 0)
-		D_GOTO(out_svc, rc);
-
-	ABT_rwlock_rdlock(svc->ps_lock);
-
-	prop_bits = DAOS_PO_QUERY_PROP_ALL;
-	rc        = pool_prop_read(&tx, svc, prop_bits, &prop);
-	if (rc != 0) {
-		D_ERROR(DF_UUID ": cannot get access data for pool, "
-				"rc=" DF_RC "\n",
-			DP_UUID(in->pcai_op.pi_uuid), DP_RC(rc));
-		D_GOTO(out_lock, rc);
-	}
-	D_ASSERT(prop != NULL);
-
-	rc = pool_check_cred_access(&in->pcai_cred, in->pcai_flags, prop);
-	if (rc != 0) {
-		D_INFO(DF_UUID ": rejecting access based on credential, "
-			       "rc=" DF_RC "\n",
-		       DP_UUID(in->pcai_op.pi_uuid), DP_RC(rc));
-	}
-
-	out->pcao_op.po_map_version = ds_pool_get_version(svc->ps_pool);
-out_lock:
-	ABT_rwlock_unlock(svc->ps_lock);
-	rdb_tx_end(&tx);
-	if (prop)
-		daos_prop_free(prop);
-out_svc:
-	ds_rsvc_set_hint(&svc->ps_rsvc, &out->pcao_op.po_hint);
-	pool_svc_put_leader(svc);
-out:
-	out->pcao_op.po_rc = rc;
-	crt_reply_send(rpc);
-}
-
 /* Currently we only maintain compatibility between 2 versions */
 #define NUM_POOL_VERSIONS	2
